refactor : RacingCarController 생성 시 NumberGenerator의 구현체를 주입받도록 리팩토링
