+++ src/main/java/racingcar/util/NumberGenerator.java
@@ -1,0 +1,6 @@
+package racingcar.util;
+
+public interface NumberGenerator {
+
+    int generate();
+}
