+++ src/main/java/racingcar/util/RandomNumberGenerator.java
@@ -1,12 +1,13 @@
-package racingcar;
+package racingcar.util;
 
 import java.util.Random;
 
-public class RandomNumberGenerator {
+public class RandomNumberGenerator implements NumberGenerator {
 
     private static final int MAX_RANDOM_INT = 10;
 
-    public static int generate() {
+    @Override
+    public int generate() {
         Random random = new Random();
         return random.nextInt(MAX_RANDOM_INT);
     }
